Refactor common logic from ProductCollectionViewCell & ProductDetailController to protocol.
